requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Werkzeug>=2.3.0
gunicorn>=20.0.0 